Adding a file MapManager for the occupation grid
Pretty self explanatory. This file contains the occupation grid for simulating rvis and slam navigation of the robot on your personal Unity machine.
The function UpdateMap() is still yet to be written. Whoever writes this must use data from the file LaserScanSensor to get readings for the initial position of the robot and what it sees through the LIDAR sim.