Add files via upload
1. Create empty object and name it e.g. PathFindingSystem. To this object attach the scripts GridManager.cs and Pathfinding.cs.
2. Create a 3d cube and name it e.g. EndPoint.
3. Create a 3d cube and name it Player. To this object attack the PathFollower.cs script. Drag the Endpoint object into the target variable of this script.
4. All Obstacles should be under the same layer, e.g. Obstacle. This layer then needs to be given to the GridManager and Pathfinding scripts as a variable.

Now for extra notes: When running the game it should move in the shortest path to the endpoint, diagonally as well. If you press 'Scene' during running you can see a blue path, which signifies the shortest path and there are also green grids that show the A* pathfinding grids. The red grids should be under the obstacles.
There are variables in the GridManager script like Grid world size and node radius. These need to be tweaked based on the scene itself. I used a 50 x 50 with node radius 0.5 for a plane of 10x10 in my personal scene.